backend/requirements.txt
anyio==4.0.0
asttokens==2.4.0
asyncpg==0.28.0
Babel==2.12.1
certifi==2023.7.22
click==8.1.7
devtools==0.12.2
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.109.2
greenlet==3.0.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pydantic==2.6.3
Pygments==2.16.1
python-jose==3.3.0
python-multipart==0.0.6
pyupgrade==3.3.1
PyYAML==6.0.1
rfc3986==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
tokenize-rt==4.2.1
typing_extensions==4.8.0
uvicorn==0.23.2
fastapi-socketio==0.0.10
pusher==3.3.2
pillow==10.3.0
pydantic-settings==2.2.1
